+++ .changeset/web-extract-attachment-upload.md
@@ -1,0 +1,5 @@
+---
+"@moonshot-ai/kimi-code": patch
+---
+
+Extract the composer's image/video attachment handling into a reusable composable.
